pyproject.toml
[project]
name = "slop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.3",
    "rich>=13.7.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.1",
    "openai>=1.37.0",
    "elevenlabs>=1.6.1",
    "Pillow>=10.3.0",
    "requests>=2.32.3",
    "tenacity>=8.3.0",
]

[project.scripts]
slop = "slop.cli:app"

[tool.uv]
package = true
dev-dependencies = []

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"
